finance engine stable

+++ src/routes/internal.js
@@ -967,7 +967,7 @@ SELECT
 le.id,
 le.wallet_id,
 le.direction,
-le.amount_cents AS amount,
+le.amount_cents::int AS amount,
 le.reference_type,
 le.reference_id,
 le.created_at
@@ -1662,7 +1662,7 @@ SELECT
 le.id,
 le.wallet_id,
 le.direction,
-le.amount_cents AS amount,
+le.amount_cents::int AS amount,
 le.reference_type,
 le.reference_id,
 le.created_at
